+++ src/edu/hitsz/factory/MobFactory.java
@@ -19,10 +19,10 @@ public class MobFactory implements EnemyFactory {
     private String type = "mob";
 
     @Override
-    public AbstractEnemy createEnemy(int level) {
+    public AbstractEnemy createEnemy(double level) {
         locationX = (int) (Math.random() * (Main.WINDOW_WIDTH - ImageManager.MOB_ENEMY_IMAGE.getWidth()));
         locationY = (int) (Math.random() * Main.WINDOW_HEIGHT * 0.2);
-        int val = speedY + (level - 1);
+        int val = speedY +(int)(level - 1);
         return new MobEnemy(locationX, locationY, speedX, Math.min(val, 15), hp, score, type);
     }
 }
